Fix options controlled<->uncontrolled comp switch
Some sub components (usually Compression) of the Options component
switched between controlled and uncontrolled components during the
lifetime of them. This happened because in the NewStack constructor the
initial setting of options was done in a wrong way, resulting in
initially an unset default value for certain options.

Additionally this moves the `ajv` somewhere so that it only needs to be
instantiated once.