Fix stats sorting and therefore calculation of totals (#69)
Make sure the stats are sorted in the right way from the beginning on
instead of relying on the backend to deliver it in some form. We'd
anyway need to re-sort it to be correct (that's why there was the call
to `timestamps.reverse()`), so this fixes it to be correct at the very
start.

Additionally the test ensures we actually get what we want.